show black friday banner on change log widget
